Crate moonshadow_runtime

The Moonshadow Runtime.

Primary features of this runtime include:

Ethereum compatibility

Moonshadow tokenomics

Modules

api

currency
MSHD, the native token, uses 18 decimals of precision.

opaque
Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core datastructures.

Structs

AdjustmentVariable
The adjustment variable of the runtime. Higher values will cause TargetBlockFullness to change the fees more rapidly. This low value causes changes to occur slowly over time.

AnnouncementDepositBase

AnnouncementDepositFactor

BaseFilter
Returns if calls are allowed through the filter

BlockGasLimit

BlockHashCount

BlockLength
We allow for 5 MB blocks.

BlockWeights
We allow for one half second of compute with a 6 second average block time. These values are dictated by Polkadot for the parachain.

BondDuration
Reward payments and collator exit requests are delayed by 2 hours (2 * 300 * block_time)

CooloffPeriod

CouncilMaxMembers
The maximum number of council members.

CouncilMaxProposals
The maximum number of Proposlas that can be open in the council at once.

CouncilMotionDuration
The maximum amount of time (in blocks) for council members to vote on motions. Motions may end in fewer blocks if enough votes are cast to determine the result.

DealWithFees

DefaultBlocksPerRound
Default BlocksPerRound is every hour (300 * 12 second block times)

DefaultCollatorCommission
Default fixed percent a collator takes off the top of due rewards is 20%

DefaultParachainBondReservePercent
Default percent of inflation set aside for parachain bond every round

DepositAmount

EnactmentPeriod

ExistentialDeposit

FastTrackVotingPeriod

FixedGasPrice

GenesisConfig

InflationInfo

InitializationPayment

Initialized

InstantAllowed

LaunchPeriod

MaxApprovals

MaxCollatorsPerNominator
Maximum 25 collators per nominator

MaxInitContributorsBatchSizes

MaxLocks

MaxNominatorsPerCollator
Maximum 10 nominators per collator

MaxPending

MaxProposals

MaxProxies

MaxReserves

MaxScheduledPerBlock

MaxVotes

MaximumSchedulerWeight

MinBlocksPerRound
Minimum round length is 2 minutes (10 * 12 second block times)

MinCollatorCandidateStk
Minimum stake required to be reserved to be a candidate is 1_000

MinCollatorStk
Minimum stake required to become a collator is 1_000

MinNominatorStk
Minimum stake required to be reserved to be a nominator is 5

MinSelectedCandidates
Minimum 8 collators selected per round, default at genesis and minimum forever after

MinimumDeposit

MinimumMultiplier
Minimum amount of the multiplier. This value cannot be too low. A test case should ensure that combined with AdjustmentVariable, we can recover from the minimum. See multiplier_can_grow_from_zero in integration_tests.rs. This value is currently only used by pallet-transaction-payment as an assertion that the next multiplier is always > min value.

MinimumPeriod

MinimumReward

MoonbeamGasWeightMapping

Origin

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PreimageByteDeposit

ProposalBond

ProposalBondMinimum

ProxyDepositBase

ProxyDepositFactor

Range

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

SS58Prefix

SpendPeriod

TargetBlockFullness
The portion of the NORMAL_DISPATCH_RATIO that we adjust the fees with. Blocks filled less than this will decrease the weight and more will increase.

TechComitteeMaxMembers
The maximum number of technical committee members.

TechComitteeMaxProposals
The maximum number of Proposlas that can be open in the technical committee at once.

TechComitteeMotionDuration
The maximum amount of time (in blocks) for technical committee members to vote on motions. Motions may end in fewer blocks if enough votes are cast to determine the result.

TransactionByteFee

TransactionConverter

TreasuryId

Version

VestingPeriod

VotingPeriod

Enums

Call

Event

OriginCaller

ProxyType
The type used to represent the kinds of proxying allowed.

Constants

DAYS

GAS_PER_SECOND
Current approximation of the gas/s consumption considering EVM execution over compiled WASM (on 4.4Ghz CPU). Given the 500ms Weight, from which 75% only are used for transactions, the total EVM execution gas limit is: GAS_PER_SECOND * 0.500 * 0.75 ~= 15_000_000.

HOURS

MAXIMUM_BLOCK_WEIGHT
Maximum weight per block

MILLISECS_PER_BLOCK

MINUTES

VERSION
This runtime version. The spec_version is composed of 2x2 digits. The first 2 digits represent major changes that can’t be skipped, such as data migration upgrades. The last 2 digits represent minor changes which can be skipped.

WASM_BINARY

WASM_BINARY_BLOATY

WEEKS

WEIGHT_PER_GAS
Approximate ratio of the amount of Weight per Gas. u64 works for approximations because Weight is a very small unit compared to gas.

Traits

BuildStorage
Complex storage builder stuff.

Functions

native_version
The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId
Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex
The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

Address
The address format for describing accounts.

AllModules	Deprecated
All modules included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystem	Deprecated
All modules included in the runtime as a nested tuple of types.

AllPallets
All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithSystem
All pallets included in the runtime as a nested tuple of types.

AuthorFilter

AuthorFilterConfig

AuthorInherent

AuthorMapping

AuthorMappingConfig

Balance
Balance of an account.

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

BlockNumber
An index to a block.

CheckedExtrinsic
Extrinsic type that has already been checked.

CouncilCollective

CouncilCollectiveConfig

CrowdloanRewards

CrowdloanRewardsConfig

Democracy

DemocracyConfig

DigestItem
Digest item type.

EVM

EVMConfig

Ethereum

EthereumChainId

EthereumChainIdConfig

EthereumConfig

Executive
Executive: handles dispatch to the various pallets.

Hash
A hash of some data used by the chain.

Header
Block header type as expected by this runtime.

Index
Index of a transaction in the chain.

ParachainInfo

ParachainInfoConfig

ParachainStaking

ParachainStakingConfig

ParachainSystem

Precompiles

Proxy

RandomnessCollectiveFlip

Scheduler

SchedulerConfig

Signature
Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock
A Block signed with a Justification

SignedExtra
The SignedExtension to the basic transaction logic.

SlowAdjustingFeeUpdate
Parameterized slow adjusting fee updated based on https://w3f-research.readthedocs.io/en/latest/polkadot/overview/2-token-economics.html#-2.-slow-adjusting-mechanism // editorconfig-checker-disable-line

Sudo

SudoConfig

System

SystemConfig

TechComitteeCollective

TechComitteeCollectiveConfig

Timestamp

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.

Utility